Do not rebuild existing site by default; provide option
Restore the behavior of the previous released version (0.4.1)
and do not always generate the Jekyll site at startup, because
this fails on read-only file systems.

By default, the site is now only generated when the destination path
is empty. (This reverts a behavior change introduced with 1088b1b.)

Also provide a configuration option to force a build at startup.