Fix broken configuration loading
Let Jekyll handle most of the configuration loading and
make sure that @path always equals the destination setting.

This fixes a bug concerning the setting of a custom destination
as initialization option: the Jekyll site used the override value
but @path was set to a different value, either the setting from
the config file, when present, or to "_site".
In consequence, the files could not be served.