Replace usage of Queue for thread communication.
This patch uses the Event pattern described in
https://emptysqua.re/blog/an-event-synchronization-primitive-for-ruby/
to tell Rack whether or not to display the wait page.  Previously, the
builder thread was pushing some data into a Queue to indicate it had
started while the Rack thread would poll the queue to see if it was
empty or not.  While this works, it's not really the way threads are
meant to alert each other about state changes.